refactor: extract shifted hex encode/decode to shared utils
